graph, store: Remove deployment_schemas.state and refactor grafting
We need a way to check whether grafting copied the data already that is
included with a deployment's metadata. To that end, we do not set the
deployment's block pointer until after we have copied from the base
deployment.

Using where the deployment's block pointer is to indicate whether a
possible graft has already been done also makes the
deployment_schemas.state field unnecessary, and makes it so that starting a
subgraph only needs to touch the shard that contains the deployment.